cordapp rewrite yo&whistleblower
