+++ whistleblower/gradle.properties
@@ -1,4 +1,4 @@
 name=Test
-group=com.template
+group=net.corda.examples.whistleblower
 version=0.1
 kotlin.incremental=false
