+++ yo-cordapp/TRADEMARK
@@ -1,4 +1,4 @@
-Corda and the Corda logo are trademarks of R3CEV LLC and its affiliates.
-All rights reserved.
+Corda and the Corda logo are trademarks of R3CEV LLC and its affiliates. All rights reserved.
 
-For R3CEV LLC's trademark and logo usage information, please consult our Trademark Usage Policy available at https://www.r3.com/trademark-policy/
+For R3CEV LLC's trademark and logo usage information, please consult our Trademark Usage Policy at
+https://www.r3.com/trademark-policy/.
