--- yo-cordapp/config/dev/log4j2-test.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="info">
-    <Properties>
-        <Property name="log-path">build/logs</Property>
-        <Property name="log-name">yo-${hostName}</Property>
-    </Properties>
-
-    <Appenders>
-        <Console name="Console-Appender" target="SYSTEM_OUT">
-            <PatternLayout>
-                <pattern>
-                    [%-5level] %d{HH:mm:ss.SSS} [%t] %c{1}.%M - %msg%n
-                </pattern>>
-            </PatternLayout>
-        </Console>
-        <File name="File-Appender"
-              fileName="${log-path}/${log-name}.log" immediateFlush="false" append="true">
-            <PatternLayout pattern="[%-5level] %d{ISO8601}{GMT+0} [%t] %c{1} - %msg%n"/>
-        </File>
-    </Appenders>
-    <Loggers>
-        <Root level="info">
-            <AppenderRef ref="Console-Appender"/>
-            <AppenderRef ref="File-Appender"/>
-        </Root>
-        <Logger name="net.corda" level="info" additivity="false">
-            <AppenderRef ref="Console-Appender"/>
-            <AppenderRef ref="File-Appender"/>
-        </Logger>
-    </Loggers>
-</Configuration>
