+++ yo-cordapp/gradle.properties
@@ -1,4 +1,4 @@
-name = Test
-group = com.template
-version = 0.1
+name=Test
+group=net.corda.examples.yo
+version=0.1
 kotlin.incremental=false
