+++ yo-cordapp/gradle/wrapper/gradle-wrapper.properties
@@ -1,3 +1,4 @@
+#Fri Aug 25 12:50:39 BST 2017
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
