+++ yo-cordapp/settings.gradle
@@ -1,1 +1,2 @@
-rootProject.name = 'yo'
+include 'workflows'
+include 'contracts'
